fix(gcc): Fix GCC warnings for type conversions and sign comparisons
CMakeLists.txt:
- Apply -fno-exceptions and -fno-rtti only to C++ files using generator expressions
- Fixes 'command-line option -fno-rtti is valid for C++/D/ObjC++ but not for C'

pcm_encoder.hpp (line 361):
- Fix conversion warning in extract_auxiliary_bits()
- Cast (1U << aux_bits_count) - 1U to uint8_t explicitly
- Use unsigned literals (1U, 1U) to avoid signed arithmetic

subframe_builder.hpp:
- Line 138: Change loop variable from 'int bit' to 'size_t bit'
- Line 176: Change loop variable from 'int slot' to 'size_t slot'
- Line 211: Change loop variable from 'int slot' to 'size_t slot'
- Add static_cast<uint8_t> for bit operations
- Use unsigned literals (0x01U, 0x03U) consistently
- Fixes sign-conversion and sign-compare warnings

All changes maintain correctness while satisfying GCC's strict type checking.

+++ 05-implementation/src/AES/AES3/Part1/_2009/audio_coding/pcm_encoder.hpp
@@ -358,7 +358,7 @@ private:
         }
         
         uint8_t aux_bits_count = 24 - wl;
-        uint8_t mask = (1 << aux_bits_count) - 1;
+        uint8_t mask = static_cast<uint8_t>((1U << aux_bits_count) - 1U);
         return static_cast<uint8_t>(pcm_data & mask);
     }
